requirements.txt
aiohttp==3.8.3
py-cord==2.2.2
Flask==2.2.2
praw==7.6.0
python-dotenv==0.21.0
